Refactor ToNative methods
