+++ speckle_connector/src/speckle_objects/geometry/length.rb
@@ -7,6 +7,10 @@ module SpeckleConnector
       def self.length_to_speckle(length, units)
         length.__send__("to_#{SpeckleConnector::Converters::SKETCHUP_UNIT_STRINGS[units]}")
       end
+
+      def self.length_to_native(length, units)
+        length.__send__(SpeckleConnector::Converters::SKETCHUP_UNIT_STRINGS[units])
+      end
     end
   end
 end
