+++ speckle_connector/src/speckle_objects/other/block_definition.rb
@@ -50,6 +50,21 @@ module SpeckleConnector
           )
         end
 
+        # finds or creates a component definition from the geometry and the given name
+        # rubocop:disable Metrics/CyclomaticComplexity
+        def self.to_native(sketchup_model, geometry, name, application_id = '', &convert)
+          definition = sketchup_model.definitions[name]
+          return definition if definition && (definition.name == name || definition.guid == application_id)
+
+          definition&.entities&.clear!
+          definition ||= sketchup_model.definitions.add(name)
+          geometry.each { |obj| convert.call(obj, definition.entities) }
+          puts("definition finished: #{name} (#{application_id})")
+          # puts("    entity count: #{definition.entities.count}")
+          definition
+        end
+        # rubocop:enable Metrics/CyclomaticComplexity
+
         def self.group_mesh_to_speckle(definition, units, definitions)
           # {material_id => Mesh}
           mat_groups = {}
